pyproject.toml
[tool.poetry]
name = "mlops-tools"
version = "0.1.0"
description = "Training project within the MLOps course"
authors = ["destitutiones <albina.rukhadze@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.0"
scikit-learn = "^1.3.1"
pandas = "^2.1.1"
matplotlib = "^3.8.0"
sphinx = ">=7.2.6"
sphinx_rtd_theme = ">=1.3.0"
hydra-core = "^1.3.2"
pathlib = "^1.0.1"
dvc = {extras = ["gdrive"], version = "^3.30.3"}
dvc-gdrive = "^2.20.0"
mlflow = "^2.8.1"
catboost = "^1.2.2"

[tool.poetry.dev-dependencies]
pre-commit = "^2.16.0"

[tool.black]
line-length = 90
target-version = ["py310"]

[tool.isort]
src_paths = ["mlops_tools"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.nbqa.config]
black = "pyproject.toml"
isort = "pyproject.toml"
flake8 = "setup.cfg"

[tool.nbqa.addopts]
flake8 = ["--extend-ignore=E402"]

[tool.nbqa.mutate]
black = 1
isort = 1

[tool.poetry.extras]
docs = ["sphinx", "sphinx_rtd_theme"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
